requirements.txt
anyio==3.6.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
exceptiongroup==1.0.1
fastapi==0.86.0
h11==0.14.0
httptools==0.5.0
idna==3.4
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
pydantic==1.10.2
PyJWT==2.6.0
pyparsing==3.0.9
pytest==7.2.0
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
sniffio==1.3.0
starlette==0.20.4
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
